extract indexed flat from zio core
